pyproject.toml
[project]
name = "Amsha"
version = "2.0.9"
description = "Amsha is a lightweight library designed to provide common configuration and helper utilities for CrewAI orchestration. It serves as a foundational component for building scalable and maintainable AI agent systems"
dependencies = [
     "PyYAML ==6.0.3",
    "crewai == 0.201.1",
    "crewai-tools == 0.75.0",
    "dependency-injector == 4.48.2",
     "pandas == 2.3.2",
    "openpyxl == 3.1.5",
    "pymongo == 4.11.3",
     "pydantic == 2.11.9",
    "docling == 2.53.0",
    "psutil == 7.1.3",
    "chardet == 5.2.0",
    "nvidia-ml-py == 13.580.82",
    "hypothesis == 6.148.7",
    "pytest == 9.0.2"
]

[tool.setuptools]
package-dir = { "" = "src/nikhil" }

[tool.setuptools.packages.find]
where = ["src/nikhil"]